Change dice animation and clean up some code

+++ CODE/chess-clock/definitions.h
@@ -77,8 +77,8 @@ struct GameState
     unsigned long time_limit = THIRTY_MINUTES;  // In deciseconds
     bool show_deciseconds = false;
 
-    short dice_number;
-    short dice[2];
+    int dice_number;
+    int dice[2];
 
     Menu current_menu = Menu::Mode;
     GameMode game_mode = GameMode::TwoClockUp;
